requirements.txt
# ============================================================
# ICE Studio — Core Runtime
# ============================================================
rich>=13.0.0
pydantic>=2.0.0
orjson>=3.9.0
python-dateutil>=2.8.0
typing-extensions>=4.8.0
pyyaml>=6.0.0

# File system & events
watchdog>=3.0.0

# CLI & TUI
click>=8.1.0
typer>=0.9.0
prompt_toolkit>=3.0.0

# Execution & process management
requests>=2.31.0
python-dotenv>=1.0.0

# Async IO / WebSockets (backend ↔ GUI)
websockets>=12.0.0

# ============================================================
# Database / Analytics Layer
# ============================================================
sqlalchemy>=2.0.0
duckdb>=1.0.0
pandas>=2.0.0
pydantic
# Optional relational database
mysql-connector-python>=8.0.0

# ============================================================
# Machine Learning Core
# ============================================================
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Embeddings / Sentence Transformers
sentence-transformers>=2.2.0

# Local LLM inference backends
ctransformers>=0.2.27

# ============================================================
# Vector Search Backends
# ============================================================
faiss-cpu>=1.7.0; sys_platform != 'win32'

# ============================================================
# Cloud LLM API (optional)
# ============================================================
openai>=1.0.0

# ============================================================
# Development Tools
# ============================================================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.1.0
mypy>=1.5.0
